Refactor File Names and Paths
- Create new directories and renamed existing directories
- Organize project files in sub-directories
- Refactor tests to correspond with specific classes
- Organize test files in sub-directories
- Perform heavy formatting
- Move test doubles into test directories
