+++ HTTPServerProject.Tests/HTTPServerResponseTests/ResponseByPathTests.cs
@@ -1,3 +1,7 @@
+using HTTPServerProject.Tests.WriteStream;
+using HTTPServerProject.Path;
+using HTTPServerProject.Tests.Parameters;
+
 namespace HTTPServerProject.ResponseByPathTests;
 
 public class UnitTestsForResponsesByPath
