+++ HTTPServerProject.Tests/HTTPServerWriteTests/TestWriteStreams.cs
@@ -1,0 +1,46 @@
+using HTTPServerProject.WriteStream;
+
+namespace HTTPServerProject.Tests.WriteStream
+{
+    public class TestWriteStreams : IWriteStreams
+    {
+        List<string> sArr = new List<string>();
+
+        public TestWriteStreams(List<string> arr)
+        {
+            sArr = arr;
+        }
+
+        public void Write(string str = default!)
+        {
+            if (str == default!)
+            {
+                sArr.Add("");
+            }
+            else
+            {
+                sArr.Add(str);
+            }
+        }
+
+        public void WriteLine(string str = default!)
+        {
+            if (str == default!)
+            {
+                sArr.Add("");
+            }
+            else
+            {
+                sArr.Add(str);
+            }
+        }
+
+        public void Flush()
+        {
+        }
+
+        public void Close()
+        {
+        }
+    }
+}
