+++ HTTPServerProject/HTTPServerProxy/ProxyClient.cs
@@ -11,6 +11,7 @@ public interface IProxyClient
 public class ProxyClient : IProxyClient
 {
     TcpClient client = new TcpClient("127.0.0.1", 8000);
+    
     public ProxyClient()
     {
     }
@@ -18,18 +19,5 @@ public class ProxyClient : IProxyClient
     public NetworkStream GetStream()
     {
         return client.GetStream();
-    }
-}
-
-public class TestProxyClient : IProxyClient
-{
-    List<string> stream = new List<string>();
-    public TestProxyClient()
-    {
-    }
-
-    public List<string> GetStream()
-    {
-        return stream;
     }
 }
